+++ internal/il/expressions.go
@@ -141,7 +141,7 @@ func (c *Compiler) VisitFieldAccess(v expressions.FieldAccessExpr) (cmd string) 
 func (c *Compiler) VisitFunctionCall(f expressions.FunctionCallExpr) (cmd string) {
 	for j, arg := range f.Arguments {
 		cmd += arg.Accept(c)
-		argName := c.functions[f.Name.Lexeme].Args[j].Name
+		argName := c.functionDefinitions[f.Name.Lexeme].Args[j].Name
 		cmd += c.CopyArg(RX, f.Name.Lexeme, argName)
 	}
 	cmd += c.Call(f.Name.Lexeme)
